+ Explicitly write temporary values to a scratch buffer which is wiped once per XTS enc/decryption rather than tracking more registers in intrinsics than exist and letting compiler spill
+ In parallel with 1 round of AES for 8 blocks, compute 1 XOR of tweak by final round key, and 1 alpha multiplication in scalar registers (should all be able to fit in execution unit gaps)

Related work items: #31475454

+++ lib/aes-neon.c
@@ -1290,7 +1290,7 @@ SymCryptXtsAesDecryptDataUnitNeon(
     keyPtr = (const __n128 *)&pExpandedKey->RoundKey[0]; \
     keyLimit = (const __n128 *)pExpandedKey->lastEncRoundKey; \
     __n128 t0, t1, r0, r0x; \
-    int aesEncryptGhashLoop; \
+    SIZE_T aesEncryptGhashLoop; \
 \
     /* Do gHashRounds full rounds (AES-128|AES-192|AES-256) with stitched GHASH */ \
     roundKey = *keyPtr++; \
@@ -1374,7 +1374,7 @@ SymCryptXtsAesDecryptDataUnitNeon(
     keyPtr = (const __n128 *)&pExpandedKey->RoundKey[0]; \
     keyLimit = (const __n128 *)pExpandedKey->lastEncRoundKey; \
     __n128 t0, t1, r0, r0x; \
-    int aesEncryptGhashLoop; \
+    SIZE_T aesEncryptGhashLoop; \
 \
     /* Do gHashRounds full rounds (AES-128|AES-192|AES-256) with stitched GHASH */ \
     roundKey = *keyPtr++; \
